Create 3 module as the project Structure for this - base-android - core - features - models
The project is a multiple Gradle module known as a multi-module project. As you can see we have a future Package that includes a module.

Every module has its own network and data layer **data-module** that has a shared domain among the app like another normal app,
at first and then will change the base URL for the endpoint in run time **domain-module** (for the use case as we use a clean Architecture).

- app
- base-android
- core
- features
- models